% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeatherStations.R
\name{getWeatherStations}
\alias{getWeatherStations}
\title{Get data frame of weather stations in Indiana}
\usage{

  getWeatherStations(isd_history_file = "ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv",
  start_year = 1999, end_year = as.numeric(format(Sys.Date(), "\%Y")),
  by = c("state", "radius"), states = "IN")
}
\description{
Get data frame of weather stations in Indiana
}

